requirements.txt
pymysql
cryptography
python-dotenv
bcrypt
flask